requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
